ISSUE-292 cleanup test records, move idm code to separate file, fix errors
